requirements.txt

# Development
black>=22.3.0
chromadb>=0.4.15

# Vector Database & Embeddings
faiss-cpu>=1.7.2
# Core Dependencies
fastapi>=0.104.0
flake8>=4.0.1
httpx>=0.25.0
isort>=5.10.1
langchain>=0.0.200
llama-index>=0.7.0
pdfplumber>=0.5.28
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Document Processing
PyMuPDF>=1.23.7

# Testing
pytest>=7.4.3
python-docx>=0.8.11

# Environment Variables
python-dotenv>=1.0.0
python-multipart>=0.0.6
sentence-transformers>=2.2.2

# Frontend
streamlit>=1.22.0
torch>=2.0.0

# LLM & RAG
transformers>=4.30.0
uvicorn>=0.24.0
